Fix bugs

+++ plugins/tenet/util/qt/waitbox.py
@@ -86,17 +86,17 @@ class WaitBox(QtWidgets.QDialog):
             self._abort_button.clicked.connect(self._abort)
             v_layout.addWidget(self._abort_button)
 
-        v_layout.setSpacing(self._dpi_scale*3)
+        v_layout.setSpacing(int(self._dpi_scale*3))
         v_layout.setContentsMargins(
-            self._dpi_scale*5,
-            self._dpi_scale,
-            self._dpi_scale*5,
-            self._dpi_scale
+            int(self._dpi_scale*5),
+            int(self._dpi_scale),
+            int(self._dpi_scale*5),
+            int(self._dpi_scale)
         )
 
         # scale widget dimensions based on DPI
         height = self._dpi_scale * 15
-        self.setMinimumHeight(height)
+        self.setMinimumHeight(int(height))
 
         # compute the dialog layout
         self.setLayout(v_layout)
